ai_model/requirements.txt
Flask==2.0.1
tensorflow==2.6.0
opencv-python==4.5.3.56
numpy==1.21.2
pandas==1.3.3
python-dotenv==0.19.2